refractor: move borsh implementation into dedicated files
