+++ code/Cargo.toml
@@ -114,7 +114,7 @@ async-recursion    = "1.1"
 async-trait        = "0.1.88"
 axum               = "0.7"
 base64             = "0.22.0"
-borsh              = "1"
+borsh              = {version = "1", features = ["de_strict_order", "derive"]}
 bs58               = { version = "0.5.1", default-features = false }
 bytes              = { version = "1", default-features = false }
 byteorder          = "1.5"
