+++ code/crates/engine/src/lib.rs
@@ -2,6 +2,7 @@ pub mod consensus;
 pub mod host;
 pub mod network;
 pub mod node;
+mod ser;
 pub mod sync;
 pub mod util;
 pub mod wal;
