+++ code/crates/engine/src/util/streaming.rs
@@ -5,7 +5,7 @@ use bytes::Bytes;
 pub type Sequence = u64;
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct StreamId(Bytes);
+pub struct StreamId(pub(crate) Bytes);
 
 impl StreamId {
     pub fn new(bytes: Bytes) -> Self {
@@ -29,7 +29,7 @@ impl fmt::Display for StreamId {
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(
     feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
+    derive(::borsh::BorshSerialize, ::borsh::BorshDeserialize)
 )]
 pub struct StreamMessage<T> {
     /// Receivers identify streams by (sender, stream_id).
@@ -92,23 +92,5 @@ impl<T> StreamContent<T> {
 
     pub fn is_fin(&self) -> bool {
         matches!(self, Self::Fin)
-    }
-}
-
-#[cfg(feature = "borsh")]
-mod _borsh {
-    use super::*;
-
-    impl borsh::BorshSerialize for StreamId {
-        fn serialize<W: borsh::io::Write>(&self, writer: &mut W) -> borsh::io::Result<()> {
-            self.0.to_vec().serialize(writer)
-        }
-    }
-
-    impl borsh::BorshDeserialize for StreamId {
-        fn deserialize_reader<R: borsh::io::Read>(reader: &mut R) -> borsh::io::Result<Self> {
-            let bytes = Vec::<u8>::deserialize_reader(reader)?;
-            Ok(StreamId(bytes.into()))
-        }
     }
 }
