+++ code/crates/sync/src/lib.rs
@@ -15,6 +15,8 @@ pub mod scoring;
 mod macros;
 mod rpc;
 
+mod ser;
+
 #[doc(hidden)]
 pub mod handle;
 pub use handle::Input;
