+++ code/crates/sync/src/ser.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "borsh")]
+mod borsh;
